Use long double for all float values.

+++ includes/complex.h
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/08 20:13:23 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/02/17 15:10:21 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/02/20 02:09:23 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@
 
 typedef struct s_complex
 {
-	double		real;
-	double		imaginary;
+	long double		real;
+	long double		imaginary;
 }				t_complex;
 
 typedef struct s_fractal
@@ -27,20 +27,20 @@ typedef struct s_fractal
 	int			iterations;
 }				t_fractal;
 
-t_complex		complex(double real, double imaginary);
+t_complex		complex(long double real, long double imaginary);
 
-double			magnitude(t_complex cmplx);
-double			quick_magnitude(t_complex cmplx);
+long double		magnitude(t_complex cmplx);
+long double		quick_magnitude(t_complex cmplx);
 t_complex		squared(t_complex cmplx);
 
 t_complex		plus(t_complex a, t_complex b);
 t_complex		times(t_complex a, t_complex b);
 
 t_fractal		mandelbrot(t_complex number, int max_iterations,
-					double infinity);
+					long double infinity);
 
 t_fractal		julia(t_complex number, t_complex factor,
 					int max_iterations,
-					double infinity);
+					long double infinity);
 
 #endif
